Moved Snapshot Result Blob to View conversion to a separate Util class. Added Tie Breaking info to barebones printable view.
