+++ resources/views/printableResults.blade.php
@@ -41,6 +41,41 @@
                     </span>
                 </div>
             @endforeach
+        <h2>Tie Breaking</h2>
+        <dl>
+            <dt>Partially-Ordered Tie-Breaking Ballot</dt>
+            <dd>
+                <span class="candidateRanking">
+                   @foreach($partiallyOrderedTieBreaker as $candidateList)
+                        @foreach($candidateList as $candidate)
+                            <span class="candidateName">{{ $candidate['name'] }}</span>
+                            @if(!$loop->last)
+                                <span class="candidateSeparator">&nbsp;=&nbsp;</span>
+                            @endif
+                        @endforeach
+                        @if(!$loop->last)
+                            <span class="candidateSeparator">&nbsp;&gt;&nbsp;</span>
+                        @endif
+                    @endforeach
+                </span>
+            </dd>
+            <dt>Totally-Ordered Tie-Breaking Ballot</dt>
+            <dd>
+                <span class="candidateRanking">
+                   @foreach($totallyOrderedTieBreaker as $candidateList)
+                        @foreach($candidateList as $candidate)
+                            <span class="candidateName">{{ $candidate['name'] }}</span>
+                            @if(!$loop->last)
+                                <span class="candidateSeparator">&nbsp;=&nbsp;</span>
+                            @endif
+                        @endforeach
+                        @if(!$loop->last)
+                            <span class="candidateSeparator">&nbsp;&gt;&nbsp;</span>
+                        @endif
+                    @endforeach
+                </span>
+            </dd>
+ 
         <h2>Results</h2>
             <span class="candidateRanking">
                @foreach($result as $candidateList)
